auth to routes

+++ block-BNaaem/bookStoreAPIsWithJwt/controllers/bookController.js
@@ -22,11 +22,17 @@ exports.getSingleBook = async ( req, res, next ) => {
 exports.createBook = async ( req, res, next ) => {
     try {
         const book = req.body;
-        if ( req.body.tags ) {
-            book.tags = req.body.tags.split(",");
+        book.userId = req.users.userId;
+        if ( req.body.title != "" ) {
+            if ( req.body.tags ) {
+                book.tags = req.body.tags.split(",");
+            }
+            if ( req.body.category ) {
+                book.category = req.body.category.charAt(0).toUpperCase()+req.body.category.slice(1);
+            }
+            const bookCreated = await Book.create( book );
+            res.status( 200 ).json( { bookCreated } );
         }
-        const bookCreated = await Book.create( book );
-        res.status( 200 ).json( { bookCreated } );
     } catch ( err ) {
         return next ( err );
     }
@@ -36,8 +42,15 @@ exports.updateBook = async ( req, res, next ) => {
     try {
         const id = req.params.id;
         const bookUpdate = req.body
-        const bookUpdated = await Book.findByIdAndUpdate( id, bookUpdate , { new: true } );
-        res.status( 200 ).json( { bookUpdated } );
+        const { userId } = await Book.findById( id );
+        if ( bookUpdate.title != "" ) {
+            if ( req.users.userId == userId  ){
+                const bookUpdated = await Book.findByIdAndUpdate( id, bookUpdate , { new: true } );
+                res.status( 200 ).json( { bookUpdated } );
+            } else {
+                res.status( 200 ).json( { error: `This book is not created by this user with email: ${req.users.email}` } )
+            }
+        }
     } catch ( err ) {
         return next ( err )
     }
@@ -46,9 +59,19 @@ exports.updateBook = async ( req, res, next ) => {
 exports.deleteBook = async ( req, res, next ) => {
     try {
         const id = req.params.id;
-        const bookDeleted = await Book.findByIdAndDelete( id );
-        const comments = await Comment.deleteMany({bookId:id})
-        res.status( 200 ).json( { bookDeleted } );
+        const { userId } = await Book.findById( id );
+        if ( req.users.userId == userId  ){
+            const bookDeleted = await Book.findByIdAndDelete( id );
+            if ( bookDeleted.bookId.length >= 1 ) {
+                const comment = await Comment.deleteMany( { bookId: id } );
+                if ( comment.usersId ) {
+
+                }
+            }
+            res.status( 200 ).json( { bookDeleted } );
+        } else {
+            res.status( 200 ).json( { error: `This book is not created by this user: ${req.users.name}` } )
+        }
     } catch ( err ) {
         return next ( err );
     }
